requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==3.1.5
bottle==0.12.18
click==7.1.2
certifi==2020.6.20
chardet==3.0.4
flask>=1.1.2
future==0.18.2
gast==0.3.3
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
itsdangerous==1.1.0
idna==2.10
Jinja2>=2.11.2
line-bot-sdk==1.16.0
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.0
opencv-python==4.3.0.36
oauthlib==3.1.0
pillow==7.2.0
protobuf==3.12.4
PyYAML>=5.3.1
python-dotenv==0.14.0
requests>=2.24.0
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorflow-cpu==2.3.0
termcolor==1.1.0
urllib3>=1.25.10
Werkzeug==1.0.1